pyproject.toml
[project]
name = "llmextractor"
version = "0.1.0"
description = ""
authors = [
    {name = "NIKHIL KUMAR REDDY",email = "nikhilkumar.m@skil.ai"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "openai (>=1.63.0,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "docx2pdf (>=0.1.8,<0.2.0)",
    "instructor (>=1.7.2,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
